pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
exclude = [".venv*", "tests*", "data*", "docs*", "scripts*"]

[project]
name = "agent-test"
version = "0.1.0"
description = "LangGraph 기반 시퀀셜 멀티 에이전트 파이프라인"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
]
